# realty-web/realty-rest/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Realty REST API
  description: >-
    Read-only HTTP surface over Realty's region economy data. This service never
    migrates or writes -- every endpoint is backed by query methods only.
  version: "1.0"
servers:
  - url: /
paths:
...
  /v1/resource-pack:
    get:
      summary: The resource pack for texturing previews
      operationId: resourcePack
      description: >-
        Where to fetch the resource pack a browser-side renderer should texture
        previews with, named by the query-service module's `resource-pack-url`
        setting.


        A location only -- the URL and its advertised SHA-1 -- never the pack's bytes.
        That URL is already public, since every joining player receives it, so
        reporting it redistributes nothing; serving the pack through this service
        would mean redistributing whatever assets it contains.


        `url` is null when the operator has configured no pack, which is the default.
        That is a normal answer rather than an error, but note what it costs: without
        a pack a renderer does not draw blocks untextured, it does not draw most of
        them at all, because their block models carry no geometry. A plot then looks
        like a failed capture.


        This is deliberately not server.properties' `resource-pack`. That pack is sent
        to the game client, which applies it over its own copy of the game, so it is
        usually an override pack; a browser has no vanilla assets underneath it. Name a
        pack that stands on its own.


        Answered entirely by the query-service module, because the setting lives with
        the game server and not in this service's database. An unreachable module is
        therefore a 502 rather than an empty pack: "none configured" and "could not
        ask" are different answers.
      responses:
        "200":
          description: The configured pack, or nulls when none is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePackResponse"
        "502":
          description: >-
            `RESOURCE_PACK_UNAVAILABLE` -- the query-service module is disabled or
            unreachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "`INTERNAL_ERROR` -- a generic message only, never the underlying exception."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ResourcePackResponse:
      type: object
      required: [url, attribution, hash, required]
      properties:
        url:
          type: string
          nullable: true
          description: Where the pack is hosted, or null when the server sets none.
        attribution:
          type: array
          description: >-
            Credits for the pack, to render wherever its textures are used and nowhere
            else. Empty when the operator configures none.


            Configured beside the pack URL, in the query-service module, because the
            operator who picks a pack is the one who knows what its licence asks for --
            and most pack licences ask for a credit. Splitting the two across two files
            on two hosts is how one of them goes stale.
          items:
            $ref: "#/components/schemas/ResourcePackAttribution"
        hash:
          type: string
          nullable: true
          description: >-
            Reserved; always null. A browser has nothing to verify a pack against, so
            no hash is configured or reported.
        required:
          type: boolean
          description: Reserved; always false. A preview cannot compel a download.
    ResourcePackAttribution:
      type: object
      required: [text, url]
      description: One credit line for the resource pack.
      properties:
        text:
          type: string
          description: The credit, rendered as text and never as markup.
        url:
          type: string
          nullable: true
          description: >-
            Where the credit links, or null for plain text. Always an absolute http(s)
            URL -- the module rejects anything else at startup, and a client should
            re-check rather than trust it, since the value is operator-supplied and
            ends up in a page.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A stable, machine-readable error code (e.g. `WORLD_NOT_FOUND`).
        message:
          type: string
          description: A human-readable message. Never the raw text of an internal exception.
      required: [ error, message ]
...